Some code reorganization.
Split up protocol.c into server.c and worker.c
Still very rough, just split it up and fixed build errors.
