+++ src/datatypes/vector.h
@@ -11,6 +11,7 @@
 #include <math.h>
 #include "../renderer/samplers/sampler.h"
 #include "../utils/assert.h"
+#include "../includes.h"
 
 struct vector {
 	float x, y, z;
