+++ src/utils/protocol/server.h
@@ -1,16 +1,17 @@
 //
-//  protocol.h
+//  server.h
 //  C-Ray
 //
-//  Created by Valtteri Koskivuori on 21/03/2021.
+//  Created by Valtteri Koskivuori on 06/04/2021.
 //  Copyright © 2021 Valtteri Koskivuori. All rights reserved.
 //
 
 #pragma once
 
-struct renderClient;
-struct renderTile;
-struct texture;
+#ifndef WINDOWS
+#include <arpa/inet.h>
+#endif
+
 struct renderer;
 
 enum clientState {
@@ -23,10 +24,6 @@ enum clientState {
 	Rendering,
 	Finished
 };
-
-#ifndef WINDOWS
-#include <arpa/inet.h>
-#endif
 
 struct renderClient {
 #ifndef WINDOWS
@@ -41,7 +38,5 @@ struct renderClient {
 // Synchronise renderer state with clients, and return a list of clients
 // ready to do some rendering
 struct renderClient *syncWithClients(const struct renderer *r, size_t *count);
-
-int startWorkerServer(void);
 
 void *networkRenderThread(void *arg);
