+++ src/utils/protocol/worker.h
@@ -1,0 +1,11 @@
+//
+//  worker.h
+//  C-Ray
+//
+//  Created by Valtteri Koskivuori on 06/04/2021.
+//  Copyright © 2021 Valtteri Koskivuori. All rights reserved.
+//
+
+#pragma once
+
+int startWorkerServer(void);
